refactor(param-id): extract the AADC tape backend into param_id/aadc_backend.py
Pure move. paramID.py 3042 -> 2765 lines.

_aadc_cost_and_grad and its nested closures (cost_on_tape, _cost_scale,
_operand_is_state, _resolve_state_idx) move wholesale to cost_and_grad(pid, ...).
This was the cleanest cut of the three: the tape re-implements the cost by hand
so that the cost and its gradient come out of one tape evaluation, which means
the code calls no generic method of the param-id object at all -- only
pid.sim_helper and the raw obs/protocol/param dicts.

The `import aadc` stays lazy inside cost_on_tape rather than being hoisted to
module scope. Hoisting would change when ImportError surfaces on a machine
without AADC, which is a behaviour change, not a move. Asserted in verification
that importing aadc_backend does not pull in aadc.

TAPE_CONSISTENT_AADC_METHODS remains importable from paramID as a re-export.

Verified with a Matlogica licence: tests/test_param_id.py -k aadc 7 passed,
tests/test_aadc_vs_casadi_3compartment.py 13 passed / 3 skipped -- both
unchanged. Plus the standard selection, byte-identical outcomes before and after
(39 tests, diffed by node id).
